+++ CHANGELOG.md
@@ -9,10 +9,12 @@ All Notable changes to `PHP Domain Parser` **5.x** series will be documented in 
 - `Pdp\PublicSuffixListSection` interface implemented by `Pdp\Rules` and `Pdp\PublicSuffix`
 - `Pdp\DomainInterface` interface implemented by `Pdp\Domain` and `Pdp\PublicSuffix`
 - `Pdp\Domain::getContent` replaces `Pdp\Domain::getDomain`
+- `Pdp\Domain::withLabel` adds a new label to the `Pdp\Domain`.
+- `Pdp\Domain::withoutLabel` removes labels from the `Pdp\Domain`.
 - `Pdp\Domain::withPublicSuffix` updates the `Pdp\Domain` public suffix part.
 - `Pdp\Domain::withSubDomain` updates the `Pdp\Domain` sub domain part.
-- `Pdp\Domain::withLabel` adds a new label to the `Pdp\Domain`.
-- `Pdp\Domain::withoutLabels` removes labels from the `Pdp\Domain`.
+- `Pdp\Domain::append` appends a label to `Pdp\Domain`.
+- `Pdp\Domain::prepend` prepends a label to `Pdp\Domain`.
 - `Pdp\Domain::resolve` attach a public suffix to the `Pdp\Domain`.
 - `Pdp\Domain::isResolvable` tells whether the current `Pdp\Domain` can have a public suffix attached to it or not.
 - `Pdp\PublicSuffix::createFromDomain` returns a new `Pdp\PublicSuffix` object from a `Pdp\Domain`object